refactor(ui): extract transcript range projection
